Move From trait for error to api
