+++ src/unbanner/mod.rs
@@ -28,20 +28,3 @@ pub enum UnbanStatus {
     // Exec name, error
     Error(String, String),
 }
-
-impl From<Vec<UnbanStatus>> for ErrorResponse {
-    fn from(uss: Vec<UnbanStatus>) -> Self {
-        let mut desc = HashMap::new();
-        for us in uss {
-            match us {
-                UnbanStatus::Ok(s) => desc.insert(s, "OK".to_string()),
-                UnbanStatus::Error(s, c) => desc.insert(s, c),
-            };
-        }
-        ErrorResponse {
-            code: 500,
-            reason: "Some executors didn't unban successfully".to_string(),
-            details: Some(desc),
-        }
-    }
-}
